Fix status handling for VLC
VLC 2.2.0 uses a new default RC interface which
is not currently handled by Bongo.  For now, use
the oldrc interface to get the track information.

Preemptively get the information by looking for
"audio volume" events, which are printed at the
start of the program, rather than requiring the
user to seek first.