Wrap internal UI updates in `with-silent-modifications` form
Currently the buffer is marked as modified even for internal UI updates
(like adding track timings), this can be annoying since it adds irrelevant
entries the undo list of the Bongo buffer leading to unexpected behaviour when
the user does an undo. As a concrete example consider this

> User removes a track
> The timing of the currently playing track is updated (via Bongo)

Without this commit each of the above events, will each generate an undo event,
so if user will wants to undo the addition of the track he will have to execute
undo twice instead of just once.

Using `with-silent-modifications` avoids the above scenario